REM-61 Add unit tests for useBulkOperations hook with comprehensive coverage
- Implemented mocks for localization and Wails functions in useBulkOperations tests.
- Created tests for completion tracking of start and stop operations, ensuring flags reset correctly based on torrent status changes.
- Added edge case tests to verify behavior when only some torrents change status during operations.
- Included reset scenarios to check immediate flag resets when no torrents need processing or when selected torrents become empty.
- Enhanced test coverage for handling cases where selected torrents disappear mid-operation.
